improve math connection lines
